feat(xt): add xt.ErrorIs and envvar XT_NO_COLORS
We add `xt.ErrorIs` to check for wrapped errors, making `xt.KO` not needed when the error needs
to be verified.

We also add the possibility to turn of colors in output of xt using the environmental variable
`XT_NO_COLORS`. This is mostly to make testing easier, but can useful in other situations as well.