feat(cmd): add changelog command to generate changelog entries
We add the `changelog` command to generate changelog entries using the commit's type/scope
with sections in the changelog. For the type/scope we use conventionalcommits.org, and
the changelog itself keepachangelog.com.

The latest tag is gathered from Git, and next tag calculated incrementing the minor version.
When the patch-version needs to be incremented, command line flag `--hotfix` can be provided.